Conflicts:
	app/models/identification.rb
	app/views/identifications/_identification.html.erb
	app/views/observations/show.html.erb
	db/schema.rb
	spec/models/identification_spec.rb
	spec/models/observation_spec.rb

+++ app/models/life_list.rb
@@ -136,7 +136,8 @@ class LifeList < List
       Taxon.to_s, [taxon.id, taxon.ancestor_ids].flatten.compact
     ]) do |list_rule|
       next unless list_rule.list.is_a?(LifeList)
-      LifeList.delay(:priority => 1).add_taxa_from_observations(list_rule.list, :taxa => [taxon.id])
+      next if Delayed::Job.where("handler LIKE '%add_taxa_from_observations%id: ''#{list_rule.list_id}''%'").exists?
+      LifeList.delay(:priority => INTEGRITY_PRIORITY).add_taxa_from_observations(list_rule.list, :taxa => [taxon.id])
     end
   end
   
@@ -157,11 +158,11 @@ class LifeList < List
   
   def self.repair_observed(list)
     ListedTaxon.do_in_batches(
-        :include => :last_observation, 
+        :include => [{:last_observation => :taxon}, :taxon], 
         :conditions => [
           "list_id = ? AND observations.id IS NOT NULL AND observations.taxon_id != listed_taxa.taxon_id",
           list.id]) do |lt|
-      lt.destroy
+      lt.destroy unless lt.last_observation.taxon.descendant_of?(lt.taxon)
     end
   end
